user-service/requirements.txt
fastapi==0.115.12
uvicorn
SQLAlchemy==2.0.34
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.0.3
pg8000==1.31.2
psycopg2-binary==2.9.9
passlib==1.7.4
PyJWT
pydantic[email]